+++ main/app_main.h
@@ -1,7 +1,7 @@
 #ifndef __APP_MAIN_PORTFOLIO2_H
 #define __APP_MAIN_PORTFOLIO2_H
 
-uint32_t measure_temp(void);
+float measure_temp(void);
 void send_response(char *payload, int payloadSize);
 
 #endif
